// src/lib/server/schemas/rsstream.ts
export const enclosureSchema = z.object({
  photo: z.array(z.object({
    link: z.string().optional(),
    alt: z.string().optional(),
  })).optional(),
  audio: z.array(z.string()).optional(),
  video: z.array(z.string()).optional(),
}).optional()

export const addItemSchema = z.object({
  url: z.string().min(1),
  guid: z.string().min(1),
  title: z.string().optional(),
  content: z.string().optional(),
  author: z.string().optional(),
  permalink: z.string().optional(),
  timestamp: z.number().int().positive(),
  enclosure: enclosureSchema,
  category: z.array(z.string()).optional(),
})